Society browse page and refactor
